GPRをファイル分離
