+++ src/cpu/CMakeLists.txt
@@ -1,2 +1,2 @@
 add_library(cpu
-    cpu.cpp cpu.h cop0.h instruction.h)
+    cpu.cpp cpu.h cop0.h instruction.h gpr.cpp)
