+++ src/cpu/gpr.cpp
@@ -1,0 +1,8 @@
+﻿#include "gpr.h"
+
+namespace N64 {
+namespace Cpu{
+
+
+} // namespace Cpu
+} // namespace N64
